requirements.txt
python-dotenv
openai
config 
psycopg2
json
flask
